requirements.txt
playwright>=1.54
pypdf>=5.0
pypdfium2>=4.30
Pillow>=10.0
pytest>=8.0
